refactor: university 관련 Type 클래스 패키지 이동
- LanguageTestType, SemesterAvailableForDispatch, TuitionFeeType을 university.domain 으로
