+++ src/main/java/com/example/solidconnection/admin/dto/LanguageTestScoreUpdateRequest.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.admin.dto;
 
 import com.example.solidconnection.custom.validation.annotation.RejectedReasonRequired;
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import com.example.solidconnection.type.VerifyStatus;
 import jakarta.validation.constraints.NotNull;
