+++ src/main/java/com/example/solidconnection/application/domain/LanguageTest.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.application.domain;
 
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
 import jakarta.persistence.EnumType;
