+++ src/main/java/com/example/solidconnection/application/dto/ApplicantResponse.java
@@ -1,7 +1,7 @@
 package com.example.solidconnection.application.dto;
 
 import com.example.solidconnection.application.domain.Application;
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 
 public record ApplicantResponse(
         String nicknameForApply,
