+++ src/main/java/com/example/solidconnection/score/dto/LanguageTestScoreRequest.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.score.dto;
 
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
