+++ src/main/java/com/example/solidconnection/score/repository/LanguageTestScoreRepository.java
@@ -3,7 +3,7 @@ package com.example.solidconnection.score.repository;
 import com.example.solidconnection.score.domain.LanguageTestScore;
 import com.example.solidconnection.score.repository.custom.LanguageTestScoreFilterRepository;
 import com.example.solidconnection.siteuser.domain.SiteUser;
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
