+++ src/main/java/com/example/solidconnection/university/domain/LanguageTestType.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.type;
+package com.example.solidconnection.university.domain;
 
 import java.util.Comparator;
