+++ src/main/java/com/example/solidconnection/university/repository/LanguageRequirementRepository.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.university.repository;
 
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import com.example.solidconnection.university.domain.LanguageRequirement;
 import com.example.solidconnection.university.domain.UniversityInfoForApply;
 import org.springframework.data.jpa.repository.JpaRepository;
