+++ src/main/java/com/example/solidconnection/university/repository/custom/UniversityFilterRepository.java
@@ -1,6 +1,6 @@
 package com.example.solidconnection.university.repository.custom;
 
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import com.example.solidconnection.university.domain.University;
 import com.example.solidconnection.university.domain.UniversityInfoForApply;
