+++ src/main/java/com/example/solidconnection/university/repository/custom/UniversityFilterRepositoryImpl.java
@@ -2,7 +2,7 @@ package com.example.solidconnection.university.repository.custom;
 
 import com.example.solidconnection.entity.QCountry;
 import com.example.solidconnection.entity.QRegion;
-import com.example.solidconnection.type.LanguageTestType;
+import com.example.solidconnection.university.domain.LanguageTestType;
 import com.example.solidconnection.university.domain.QUniversity;
 import com.example.solidconnection.university.domain.QUniversityInfoForApply;
 import com.example.solidconnection.university.domain.University;
